+++ CHANGELOG.md
@@ -29,6 +29,11 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 
 ---
 
+## [4.32] Added IconButton and StackedButton components
+
+- Added `IconButton` to components
+- Added `StackedButton` to components
+
 ## [4.31] Add background classes for gray and aqua.
 - Added bgColor--gray and bgColor--aqua helper classes